Refactored review/explain command with helper function
