+++ src/commands/getExplainCodeCommand.ts
@@ -1,11 +1,9 @@
 import * as vscode from "vscode";
-import { ChatRole, TextBlock } from "../ChatModel/ChatModel";
 import { ChatManager } from "../Chat/ChatManager";
-import { getCodeReference } from "../helpers/getCodeReference";
-import { sendChatMessage } from "../helpers/sendChatMessage";
 import { SettingsProvider } from "../helpers/SettingsProvider";
 import { ChatWebviewProvider } from "../providers/ChatViewProvider";
-import { ensureActiveWebviewAndChat } from "../helpers/ensureActiveWebviewAndChat";
+import { insertMessage } from "../helpers/insertMessage";
+import { TextBlock } from "../ChatModel/ChatModel";
 
 /**
  * Command to explain the selected code (or whole file if no selection) in a chat.
@@ -29,24 +27,9 @@ export const getExplainCodeCommand = (
       content: settingsProvider.getExplainCodePrompt(),
     } as TextBlock;
 
-    const codeReference = getCodeReference(activeEditor);
-
-    const message = {
-      role: ChatRole.User,
-      content: codeReference ? [textBlock, codeReference] : [textBlock],
-    };
-
-    await ensureActiveWebviewAndChat(chatManager, chatWebviewProvider);
-
-    const chat = chatManager.getActiveChat();
-    chat?.messages.push(message);
-
-    chatWebviewProvider.updateChat(chatManager.chats, chatManager.activeChatId);
-
-    chatWebviewProvider.setLoading(true);
-
-    await sendChatMessage(
-      message,
+    insertMessage(
+      textBlock,
+      activeEditor,
       settingsProvider,
       chatManager,
       chatWebviewProvider,
